Refactored Base Structure
