+++ app/src/main/java/me/morglenrazor/imdbcollector/presentaiton/MovieListView.kt
@@ -1,0 +1,6 @@
+package me.morglenrazor.imdbcollector.presentaiton
+
+import me.morglenrazor.imdbcollector.presentaiton.base.BaseView
+
+class MovieListView : BaseView {
+}
