+++ app/src/main/java/me/morglenrazor/imdbcollector/presentaiton/main/MainView.kt
@@ -1,10 +1,11 @@
 package me.morglenrazor.imdbcollector.presentaiton.main
 
 import me.morglenrazor.imdbcollector.presentaiton.base.BaseView
+import moxy.MvpView
 import moxy.viewstate.strategy.alias.AddToEndSingle
 import moxy.viewstate.strategy.alias.Skip
 
-interface MainView : BaseView {
+interface MainView : MvpView {
     @Skip
     fun setRootScreen()
 }
