+++ app/src/main/java/me/morglenrazor/imdbcollector/ui/CurrentMovieFragment.kt
@@ -1,0 +1,6 @@
+package me.morglenrazor.imdbcollector.ui
+
+import me.morglenrazor.imdbcollector.ui.base.BaseFragment
+
+class CurrentMovieFragment : BaseFragment() {
+}
