+++ app/src/main/java/me/morglenrazor/imdbcollector/ui/MovieListFragment.kt
@@ -1,4 +1,4 @@
-package me.morglenrazor.imdbcollector.ui.main
+package me.morglenrazor.imdbcollector.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,8 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import me.morglenrazor.imdbcollector.R
+import me.morglenrazor.imdbcollector.ui.base.BaseFragment
 
-class MovieListFragment : Fragment() {
+class MovieListFragment : BaseFragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
