--- app/src/main/java/me/morglenrazor/imdbcollector/ui/base/BaseActivity.kt
@@ -1,10 +1,0 @@
-package me.morglenrazor.imdbcollector.ui.base
-
-import android.os.Bundle
-import android.widget.Toast
-import me.morglenrazor.imdbcollector.presentaiton.base.BaseView
-import moxy.MvpAppCompatActivity
-
-abstract class BaseActivity : MvpAppCompatActivity(), BaseView {
-
-}
